Remove interfaces in model.git package
