+++ src/test/java/de/uhd/ifi/se/decision/management/jira/extraction/tangledchangedetector/TestTangledChangeDetector.java
@@ -14,8 +14,8 @@ import de.uhd.ifi.se.decision.management.jira.extraction.gitclient.TestSetUpGit;
 import de.uhd.ifi.se.decision.management.jira.extraction.impl.TangledChangeDetectorImpl;
 import de.uhd.ifi.se.decision.management.jira.model.git.ChangedFile;
 import de.uhd.ifi.se.decision.management.jira.model.git.Diff;
+import de.uhd.ifi.se.decision.management.jira.model.git.Diff;
 import de.uhd.ifi.se.decision.management.jira.model.git.TestDiff;
-import de.uhd.ifi.se.decision.management.jira.model.git.impl.DiffImpl;
 
 public class TestTangledChangeDetector extends TestSetUpGit {
 
@@ -91,7 +91,7 @@ public class TestTangledChangeDetector extends TestSetUpGit {
     @Test
     public void testCalculatePackageDistanceRightBiggerLeft() {
 	Diff diffForJiraIssue = TestDiff.createDiff(mockJiraIssueForGitTestsTangled);
-	Diff diff = new DiffImpl();
+	Diff diff = new Diff();
 
 	diff.addChangedFile(diffForJiraIssue.getChangedFiles().get(2));
 	diff.addChangedFile(diffForJiraIssue.getChangedFiles().get(1));
