+++ src/test/java/de/uhd/ifi/se/decision/management/jira/model/git/TestDiff.java
@@ -11,7 +11,6 @@ import org.junit.Test;
 import com.atlassian.jira.issue.Issue;
 
 import de.uhd.ifi.se.decision.management.jira.extraction.gitclient.TestSetUpGit;
-import de.uhd.ifi.se.decision.management.jira.model.git.impl.DiffImpl;
 
 public class TestDiff extends TestSetUpGit {
 
@@ -32,7 +31,7 @@ public class TestDiff extends TestSetUpGit {
 
     @Test
     public void createDiff() {
-	Diff diff = new DiffImpl();
+	Diff diff = new Diff();
 	assertEquals(0, diff.getChangedFiles().size());
     }
